requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.0.5
pathspec==0.8.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-firebase==1.2
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
